Extract updater function from Nav Link block
This will make this easier to refactor in future
